pyproject.toml
[project]
name = "vbg-gn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dash==2.17.1",
    "dash-bootstrap-components==1.6.0",
    "flask==3.0.3",
    "gunicorn==22.0.0",
    "pandas>=2.3.3",
    "plotly==5.22.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.30",
]
